# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PXE Boot Provisioning and Node Management System API.",
        "title": "OS Baka API",
        "contact": {
            "name": "OS Baka Project",
            "url": "https://github.com/os-baka"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/pxe/boot/{mac}": {
            "get": {
                "description": "Returns an iPXE script based on the node's MAC address",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "pxe"
                ],
                "summary": "Get iPXE boot script for a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC Address",
                        "name": "mac",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "iPXE script",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/pxe/postinstall/{mac}": {
            "get": {
                "description": "Returns a script to run after OS installation",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "pxe"
                ],
                "summary": "Get post-installation script",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC Address",
                        "name": "mac",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post-install script",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pxe/preseed/{mac}": {
            "get": {
                "description": "Returns a preseed file for automated installation",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "pxe"
                ],
                "summary": "Get Preseed configuration for Debian/Ubuntu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC Address",
                        "name": "mac",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preseed configuration",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}